Create bit.c
You’re given two binary matrices, m rows and n columns. Both matrices have the same size i.e. same number of rows and same number of columns. Each cell is a bit either 0 or 1. You need to find the minimum number of swaps are required in the first matrix in order for it’s arrangement to match that of the second. If the matched arrangement can never be achieved output -1.